remove the x64 cpu define fix. Only keep the fix to types

+++ public/steam/steamtypes.h
@@ -77,11 +77,11 @@ typedef long int lint64;
 typedef unsigned long int ulint64;
 
 #ifdef X64BITS
-typedef int64_t intp;
-typedef uint64_t uintp;
+typedef long long intp;
+typedef unsigned long long uintp;
 #else
-typedef int32_t intp;
-typedef uint32_t uintp;
+typedef int intp;
+typedef unsigned int uintp;
 #endif
 
 #endif // else _WIN32
